pyproject.toml
[project]
name = "python-based-temperature-acquisition-system-for-the-host-computer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyside6>=6.11.1",
    "pyserial>=3.5",
    "pyqtgraph>=0.13",
    "openpyxl>=3.1",
]
